newPost moved to Posts folder
