+++ src/config/routes.js
@@ -6,7 +6,7 @@ import UserShow from '../pages/UserShow';
 import SignUp from '../pages/SignUp';
 import LogIn from '../pages/LogIn';
 import PostList from '../pages/PostList';
-import NewPost from '../pages/NewPost';
+import NewPost from '../pages/Posts/NewPost';
 import PostDetail from '../pages/Posts/PostDetail';
 import EditPost from '../pages/Posts/EditPost';
 import EditUser from '../pages/EditUser';
